feat: migrates to null safety

+++ test/fixtures/some_folder/another_hello.dart
@@ -1,5 +1,5 @@
 import 'dart:math';
 
 double coolFunction() {
-  return pow(2, 10);
+  return pow(2, 10) as double;
 }
